Clean up imports as well as licensing and useClipboard inside of the utils
