+++ packages/ui/src/common/Accordion/AccordionItem.tsx
@@ -27,8 +27,9 @@ import { MouseEvent, useEffect, useRef } from 'react';
 
 import type { Theme } from '../../theme';
 import { useThemeContext } from '../../theme/ThemeContext';
+import { useClipboard } from '../../utils/useClipboard';
 import ReadMoreText from '../ReadMoreText';
-import { AccordionData, AccordionOpenState, useClipboard } from './Accordion';
+import { AccordionData, AccordionOpenState } from './Accordion';
 
 const MAX_LINES_BEFORE_EXPAND = 2;
 
@@ -170,7 +171,8 @@ const AccordionItem = ({ index, accordionData, openState }: AccordionItemProps) 
 	const accordionRef = useRef<HTMLLIElement>(null);
 	const theme: Theme = useThemeContext();
 	const { setClipboardContents } = useClipboard();
-	const { description, title, content, schemaName } = accordionData;
+
+	const { description, title, content } = accordionData;
 	const { ChevronDown, Hash } = theme.icons;
 
 	const indexString = index.toString();
