+++ packages/ui/src/viewer-table/InteractionPanel/ExpandAllButton.tsx
@@ -20,9 +20,10 @@
  */
 
 import React from 'react';
+
 import Button from '../../common/Button';
-import { useThemeContext } from '../../theme/ThemeContext';
 import { Theme } from '../../theme';
+import { useThemeContext } from '../../theme/ThemeContext';
 
 export interface ExpandAllButtonProps {
 	onClick: () => void;
