+++ packages/ui/src/viewer-table/OpenModalButton.tsx
@@ -20,6 +20,7 @@
 
 import { css } from '@emotion/react';
 import { ReactNode } from 'react';
+
 import { Theme } from '../theme';
 import { useThemeContext } from '../theme/ThemeContext';
 import Eye from '../theme/icons/Eye';
